pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "benchback-rl"
version = "0.1.0"
description = "Benchmarking Backends for RL - PyTorch, JAX with Flax.NNX, JAX with Flax.Linen"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT"}

authors = [
    { name = "Daniel Kluvanec"}
]

dependencies = [
    "jax>=0.7.2.dev0",
    "jaxlib>=0.7.2.dev0",
    "flax>=0.11.2",
    "optax>=0.2.6",
    "torch>=2.5.0",
    "numpy>=2.3.0",
    "gymnax>=0.0.9",
    "pydantic>=2.12.5",
    "pydantic_settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[project.urls]
"Homepage" = "https://github.com/KluvaDa/benchback-rl"

[tool.ruff]
line-length = 120

[tool.ruff.format]
docstring-code-format = true
